test: int32 batch_idx for staged Ewald bindings batched cases
The stage-2 staged reciprocal bindings launch Warp kernels that require int32
batch indices; the batched test helper built them as int64.

+++ test/models/test_ewald_staged_bindings.py
@@ -118,7 +118,9 @@ def _nacl_batched(
             torch.full(
                 (data["positions"].shape[0],),
                 s,
-                dtype=torch.int64,
+                # Stage-2 staged bindings launch Warp kernels that require int32
+                # batch indices; the monolithic reference accepts either.
+                dtype=torch.int32,
                 device=device,
             )
         )
